[Hades:] Why do we build the wall My children, my children? Why do we build the wall?
[Company:]
Why do we build the wall?
We build the wall to keep us free
That's why we build the wall
We build the wall to keep us free

[Hades:]
How does the wall keep us free
My children, my children?
How does the wall keep us free?

[Company:]
How does the wall keep us free?
The wall keeps out the enemy
And we build the wall to keep us free
That's why we build the wall
We build the wall to keep us free

[Hades:]
Who do we call the enemy
My children, my children?
Who do we call the enemy?

[Company:]
Who do we call the enemy?
The enemy is poverty
And the wall keeps out the enemy
And we build the wall to keep us free
That's why we build the wall
We build the wall to keep us free

[Hades:]
Because we have and they have not
My children, my children
Because they want what we have got

[Company:]
Because we have and they have not
Because they want what we have got
The enemy is poverty
And the wall keeps out the enemy
And we build the wall to keep us free
That's why we build the wall
We build the wall to keep us free

[Hades:]
What do we have that they should want
My children, my children?
What do we have that they should want?

[Eurydice:]
What do we have that they should want?
We have a wall to work upon
We have work and they have none

[Hades:]
And our work is never done!
My children, my children!
And the war is never won

[Hades & company:]
The enemy is poverty
And the wall keeps out the enemy
And we build the wall to keep us free
That's why we build the wall
We build the wall to keep us free
We build the wall to keep us free...